// server/src/modules/shared/infraestructure/interfaces/api-response.interface.ts
export interface ApiResponse<T> {
  success: boolean;
  code: number;
  message?: string;
  data?: T;
}

/* Example IValidationErrors:
  "password": {
            "type": "field",
            "msg": "El password es obligatorio",
            "path": "password",
            "location": "body"
        }
  */
export interface IValidationError {
  [key: string]: {
    type: string;
    msg: string;
    path: string;
    location: string;
  };
}